+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [v1.8.0] - 2022-09-14
 
 - Change default for `--origin` flag of `buf beta studio-agent` to `https://studio.buf.build`
 - Change default for `--timeout` flag of `buf beta studio-agent` to `0` (no timeout). Before it was
@@ -725,7 +725,7 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.7.0...HEAD
+[v1.8.0]: https://github.com/bufbuild/buf/compare/v1.7.0...v1.8.0
 [v1.7.0]: https://github.com/bufbuild/buf/compare/v1.6.0...v1.7.0
 [v1.6.0]: https://github.com/bufbuild/buf/compare/v1.5.0...v1.6.0
 [v1.5.0]: https://github.com/bufbuild/buf/compare/v1.4.0...v1.5.0